--- src/main/java/uk/gov/hmcts/reform/payments/client/config/ServiceTokenGeneratorConfiguration.java
@@ -1,24 +1,0 @@
-package uk.gov.hmcts.reform.payments.client.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Lazy;
-import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGeneratorFactory;
-
-@Configuration
-@Lazy
-public class ServiceTokenGeneratorConfiguration {
-
-    @Bean
-    public AuthTokenGenerator authTokenGenerator(
-        @Value("${idam.s2s-auth.totp_secret}") String secret,
-        @Value("${idam.s2s-auth.microservice}") String microService,
-        ServiceAuthorisationApi serviceAuthorisationApi
-    ) {
-        return AuthTokenGeneratorFactory.createDefaultGenerator(secret, microService, serviceAuthorisationApi);
-    }
-
-}
