+++ src/main/java/uk/gov/hmcts/reform/payments/client/functional/models/NavigationLink.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.payments.client.models;
+package uk.gov.hmcts.reform.payments.client.functional.models;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
@@ -8,6 +8,9 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import java.net.URI;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -16,7 +19,7 @@ import lombok.ToString;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class NavigationLinks {
-    private NavigationLink nextUrl;
-    private NavigationLink self;
+public class NavigationLink {
+    private URI href;
+    private RequestMethod method;
 }
